+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 * [HexaPDF::Layout::ColumnBox#fit] to correctly take initial height into account
 * [HexaPDF::Layout::ColumnBox#fit] to ensure correct results in case the
   requested dimensions are larger than the current region
+* [HexaPDF::Document::Layout#formatted_text_box] to correctly handle properties
 
 
 ## 0.34.1 - 2023-11-01